+++ lesson8/src/abstraction.ts
@@ -1,5 +1,5 @@
 import { User } from './interface-json';
-import { getJson } from './interface-json';
+
 
 abstract class AbstractUser {
     public user: User;
@@ -9,7 +9,7 @@ abstract class AbstractUser {
     public abstract getSummary(): string;
 }
 
-class DetailedUser extends AbstractUser {
+export class DetailedUser extends AbstractUser {
     public user: User;
 
     public constructor(user: User) {
@@ -22,12 +22,3 @@ class DetailedUser extends AbstractUser {
     }
 }
 
-async function main(): Promise<void> {
-    const users: User[] = await getJson();
-    users.forEach(user => {
-        const detailedUser = new DetailedUser(user);
-        console.log('Detailed User Summary:', detailedUser.getSummary());
-    });
-}
-
-main();
